Use explicit stdint.h types throughout codebase
Replaces implicit integer types with explicit fixed-width types for
better portability and clarity:

- listen_port: int → uint16_t (ports are 0-65535)
- Loop indices in protocol.c: int → size_t (array indexing)
- Error line numbers: int → int32_t (explicit width)
- Printf formats: Updated to match new types (%d → %u, PRId32)

Added includes:
- <stdint.h> to config.h
- <inttypes.h> to error.h (for PRId32 format macro)

All tests pass including sanitizer checks.

+++ src/config.c
@@ -177,7 +177,7 @@ ik_cfg_load (TALLOC_CTX *ctx, const char *path)
       json_decref (root);
       return ERR (ctx, OUT_OF_RANGE, "Port must be 1024-65535, got %lld", port_raw);
     }
-  int port_value = (int) port_raw;
+  uint16_t port_value = (uint16_t) port_raw;
 
   // Copy values to config
   cfg->openai_api_key = talloc_strdup (cfg, json_string_value (api_key));
