+++ src/config.h
@@ -1,6 +1,7 @@
 #ifndef IK_CONFIG_H
 #define IK_CONFIG_H
 
+#include <stdint.h>
 #include <talloc.h>
 #include "error.h"
 
@@ -8,7 +9,7 @@ typedef struct
 {
   char *openai_api_key;
   char *listen_address;
-  int listen_port;
+  uint16_t listen_port;
 } ik_cfg_t;
 
 ik_result_t ik_cfg_load (TALLOC_CTX * ctx, const char *path);
